@@ -1,11 +1,12 @@
 name: getflutter
 description: GetFlutter is open source libraries that come with pre-build 1000+ UI components. It makes development faster & more enjoyable. You can customize the component as per your need.
-version: 1.0.0-dev.3
+version: 1.0.5
 #author: GetFlutter <info@getflutter.dev>
 homepage: https://github.com/ionicfirebaseapp/getflutter
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  #  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
