@@ -1,12 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:getflutter/shape/gf_badge_shape.dart';
-import 'package:getflutter/size/gf_size.dart';
-import 'package:getflutter/colors/gf_color.dart';
-export 'package:getflutter/shape/gf_badge_shape.dart';
-export 'package:getflutter/size/gf_size.dart';
-export 'package:getflutter/colors/gf_color.dart';
+import 'package:getflutter/getflutter.dart';
 
 class GFBadge extends StatefulWidget {
+  /// Create badges of all types, check out [GFButtonBadge] for button badges and [GFIconBadge] for icon type badges
+  const GFBadge({
+    Key key,
+    this.textStyle,
+    this.borderShape,
+    this.shape = GFBadgeShape.standard,
+    this.color = GFColors.DANGER,
+    this.textColor = GFColors.WHITE,
+    this.size = GFSize.SMALL,
+    this.border,
+    this.text,
+    this.child,
+  })  : assert(shape != null, 'Counter shape can not be null'),
+        super(key: key);
+
   /// The border side for the badge's [Material].
   final BorderSide border;
 
@@ -16,11 +26,11 @@ class GFBadge extends StatefulWidget {
   /// Counter type of [GFBadgeShape] i.e, standard, pills, square, circle
   final GFBadgeShape shape;
 
-  /// Pass [GFColor] or [Color]
-  final dynamic color;
+  /// Pass [GFColors] or [Color]
+  final Color color;
 
   /// size of [double] or [GFSize] i.e, 1.2, small, medium, large etc.
-  final GFSize size;
+  final double size;
 
   /// child of type [Widget] is alternative to child. text will get priority over child
   final Widget child;
@@ -31,23 +41,8 @@ class GFBadge extends StatefulWidget {
   /// text style of counter text.
   final TextStyle textStyle;
 
-  /// Pass [GFColor] or [Color]
-  final dynamic textColor;
-
-  /// Create badges of all types, check out [GFButtonBadge] for button badges and [GFIconBadge] for icon type badges
-  const GFBadge({
-    Key key,
-    this.textStyle,
-    this.borderShape,
-    this.shape = GFBadgeShape.standard,
-    this.color = GFColor.danger,
-    this.textColor = GFColor.white,
-    this.size = GFSize.medium,
-    this.border,
-    this.text,
-    this.child,
-  })  : assert(shape != null, 'Counter shape can not be null'),
-        super(key: key);
+  /// Pass [GFColors] or [Color]
+  final Color textColor;
 
   @override
   _GFBadgeState createState() => _GFBadgeState();
@@ -58,18 +53,18 @@ class _GFBadgeState extends State<GFBadge> {
   Color textColor;
   Widget child;
   GFBadgeShape counterShape;
-  GFSize size;
+  double size;
   double height;
   double width;
   double fontSize;
 
   @override
   void initState() {
-    this.color = getGFColor(widget.color);
-    this.textColor = getGFColor(widget.textColor);
-    this.child = widget.text != null ? Text(widget.text ?? '') : widget.child;
-    this.counterShape = widget.shape;
-    this.size = widget.size;
+    color = widget.color;
+    textColor = widget.textColor;
+    child = widget.text != null ? Text(widget.text ?? '') : widget.child;
+    counterShape = widget.shape;
+    size = widget.size;
     super.initState();
   }
 
@@ -78,64 +73,75 @@ class _GFBadgeState extends State<GFBadge> {
     final BorderSide shapeBorder = widget.border != null
         ? widget.border
         : BorderSide(
-            color: this.color,
-            width: 0.0,
+            color: color,
+            width: 0,
           );
 
     ShapeBorder shape;
 
-    if (this.counterShape == GFBadgeShape.pills) {
+    if (counterShape == GFBadgeShape.pills) {
       shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(50.0), side: shapeBorder);
-    } else if (this.counterShape == GFBadgeShape.square) {
+        borderRadius: BorderRadius.circular(50),
+        side: shapeBorder,
+      );
+    } else if (counterShape == GFBadgeShape.square) {
       shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(0.0), side: shapeBorder);
-    } else if (this.counterShape == GFBadgeShape.standard) {
+        borderRadius: BorderRadius.circular(0),
+        side: shapeBorder,
+      );
+    } else if (counterShape == GFBadgeShape.standard) {
       shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(5.0), side: shapeBorder);
-    } else if (this.counterShape == GFBadgeShape.circle) {
+        borderRadius: BorderRadius.circular(5),
+        side: shapeBorder,
+      );
+    } else if (counterShape == GFBadgeShape.circle) {
       shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(100.0), side: shapeBorder);
+        borderRadius: BorderRadius.circular(100),
+        side: shapeBorder,
+      );
     } else {
       shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(5.0), side: shapeBorder);
+        borderRadius: BorderRadius.circular(5),
+        side: shapeBorder,
+      );
     }
 
-    if (this.size == GFSize.small) {
-      this.height = getGFSize(this.size) * 0.56;
-      this.width = getGFSize(this.size) * 0.75;
-      this.fontSize = getGFSize(this.size) * 0.31;
-    } else if (this.size == GFSize.medium) {
-      this.height = getGFSize(this.size) * 0.5;
-      this.width = getGFSize(this.size) * 0.65;
-      this.fontSize = getGFSize(this.size) * 0.3;
-      this.fontSize = 12.0;
-    } else if (this.size == GFSize.large) {
-      this.height = getGFSize(this.size) * 0.428;
-      this.width = getGFSize(this.size) * 0.535;
-      this.fontSize = getGFSize(this.size) * 0.214;
+    if (widget.size == GFSize.SMALL) {
+      height = size * 0.56;
+      width = size * 0.73;
+      fontSize = size * 0.31;
+    } else if (widget.size == GFSize.MEDIUM) {
+      height = size * 0.58;
+      width = size * 0.76;
+      fontSize = size * 0.34;
+    } else if (widget.size == GFSize.LARGE) {
+      height = size * 0.6;
+      width = size * 0.79;
+      fontSize = size * 0.37;
     } else {
-      this.height = getGFSize(this.size) * 0.56;
-      this.width = getGFSize(this.size) * 0.75;
-      this.fontSize = getGFSize(this.size) * 0.31;
+      height = size * 0.58;
+      width = size * 0.76;
+      fontSize = size * 0.34;
     }
 
     return Container(
-      height: this.height,
-      width:
-          this.counterShape == GFBadgeShape.circle ? this.height : this.width,
+      height: height,
+      width: counterShape == GFBadgeShape.circle ? height : width,
       child: Material(
-        textStyle: this.textColor != null
-            ? TextStyle(color: this.textColor, fontSize: this.fontSize)
+        textStyle: textColor != null
+            ? TextStyle(
+                color: textColor,
+                fontSize: fontSize,
+              )
             : widget.textStyle,
-        shape: widget.borderShape == null ? shape : widget.borderShape,
-        color: this.color,
+        shape: widget.borderShape ?? shape,
+        color: color,
         type: MaterialType.button,
         child: Container(
           child: Center(
-            widthFactor: 1.0,
-            heightFactor: 1.0,
-            child: this.child,
+            widthFactor: 1,
+            heightFactor: 1,
+            child: child,
           ),
         ),
       ),
