@@ -2,16 +2,40 @@ import 'dart:math' as math;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/material.dart';
-import 'package:getflutter/shape/gf_icon_button_shape.dart';
-import 'package:getflutter/size/gf_size.dart';
-import 'package:getflutter/types/gf_type.dart';
-import 'package:getflutter/colors/gf_color.dart';
-export 'package:getflutter/shape/gf_icon_button_shape.dart';
-export 'package:getflutter/size/gf_size.dart';
-export 'package:getflutter/types/gf_type.dart';
-export 'package:getflutter/colors/gf_color.dart';
+import 'package:getflutter/getflutter.dart';
 
 class GFIconButton extends StatefulWidget {
+  /// Create icon buttons of all types. check out [GFButton] for buttons
+  const GFIconButton({
+    Key key,
+    this.iconSize = 0.0,
+    this.padding = const EdgeInsets.all(8),
+    this.alignment = Alignment.center,
+    @required this.icon,
+    this.focusColor,
+    this.hoverColor,
+    this.highlightColor,
+    this.splashColor,
+    this.disabledColor,
+    @required this.onPressed,
+    this.focusNode,
+    this.autofocus = false,
+    this.tooltip,
+    this.type = GFButtonType.solid,
+    this.shape = GFIconButtonShape.standard,
+    this.color = GFColors.PRIMARY,
+    this.borderShape,
+    this.boxShadow,
+    this.size = GFSize.MEDIUM,
+    this.buttonBoxShadow,
+    this.borderSide,
+  })  : assert(iconSize != null),
+        assert(padding != null),
+        assert(alignment != null),
+        assert(autofocus != null),
+        assert(icon != null),
+        super(key: key);
+
   /// The size of the icon inside the button.
   final double iconSize;
 
@@ -30,23 +54,23 @@ class GFIconButton extends StatefulWidget {
   /// The color for the button's icon when a pointer is hovering over it.
   final Color hoverColor;
 
-  /// Button type of [GFType] i.e, solid, outline, outline2x transparent
-  final GFType type;
+  /// Button type of [GFButtonType] i.e, solid, outline, outline2x transparent
+  final GFButtonType type;
 
   /// Button type of [GFIconButtonShape] i.e, standard, pills, square, shadow, icons
   final GFIconButtonShape shape;
 
-  /// Pass [GFColor] or [Color]
-  final dynamic color;
+  /// Pass [GFColors] or [Color]
+  final Color color;
 
-  /// Pass [GFColor] or [Color]. The primary color of the button when the button is in the down (pressed) state.
-  final dynamic splashColor;
+  /// Pass [GFColors] or [Color]. The primary color of the button when the button is in the down (pressed) state.
+  final Color splashColor;
 
-  /// Pass [GFColor] or [Color]. The secondary color of the button when the button is in the down (pressed) state.
-  final dynamic highlightColor;
+  /// Pass [GFColors] or [Color]. The secondary color of the button when the button is in the down (pressed) state.
+  final Color highlightColor;
 
-  /// Pass [GFColor] or [Color]. The color to use for the icon inside the button, if the icon is disabled.
-  final dynamic disabledColor;
+  /// Pass [GFColors] or [Color]. The color to use for the icon inside the button, if the icon is disabled.
+  final Color disabledColor;
 
   /// The callback that is called when the button is tapped or otherwise activated.
   final VoidCallback onPressed;
@@ -67,44 +91,13 @@ class GFIconButton extends StatefulWidget {
   final ShapeBorder borderShape;
 
   /// size of [double] or [GFSize] i.e, 1.2, small, medium, large etc.
-  final dynamic size;
+  final double size;
 
   /// on true state default box shadow appears around button, if GFButtonType is solid
   final bool buttonBoxShadow;
 
   /// The border side for the button's [Material].
   final BorderSide borderSide;
-
-  /// Create icon buttons of all types. check out [GFButton] for buttons
-  const GFIconButton({
-    Key key,
-    this.iconSize = 0.0,
-    this.padding = const EdgeInsets.all(8.0),
-    this.alignment = Alignment.center,
-    @required this.icon,
-    this.focusColor,
-    this.hoverColor,
-    this.highlightColor,
-    this.splashColor,
-    this.disabledColor,
-    @required this.onPressed,
-    this.focusNode,
-    this.autofocus = false,
-    this.tooltip,
-    this.type = GFType.solid,
-    this.shape = GFIconButtonShape.standard,
-    this.color = GFColor.primary,
-    this.borderShape,
-    this.boxShadow,
-    this.size = GFSize.medium,
-    this.buttonBoxShadow,
-    this.borderSide,
-  })  : assert(iconSize != null),
-        assert(padding != null),
-        assert(alignment != null),
-        assert(autofocus != null),
-        assert(icon != null),
-        super(key: key);
 
   @override
   _GFIconButtonState createState() => _GFIconButtonState();
@@ -113,7 +106,7 @@ class GFIconButton extends StatefulWidget {
 class _GFIconButtonState extends State<GFIconButton> {
   Color color;
   Function onPressed;
-  GFType type;
+  GFButtonType type;
   GFIconButtonShape shape;
   BoxShadow boxShadow;
   double height;
@@ -122,49 +115,65 @@ class _GFIconButtonState extends State<GFIconButton> {
 
   @override
   void initState() {
-    this.color = getGFColor(widget.color);
-    this.onPressed = widget.onPressed;
-    this.type = widget.type;
-    this.shape = widget.shape;
+    color = widget.color;
+    onPressed = widget.onPressed;
+    type = widget.type;
+    shape = widget.shape;
     super.initState();
+  }
+
+  Color getBorderColor() {
+    if (widget.onPressed != null) {
+      return color;
+    } else {
+      if (widget.disabledColor != null) {
+        return widget.disabledColor;
+      } else {
+        return color.withOpacity(0.48);
+      }
+    }
+  }
+
+  Color getDisabledFillColor() {
+    if (widget.type == GFButtonType.transparent ||
+        widget.type == GFButtonType.outline ||
+        widget.type == GFButtonType.outline2x) {
+      return Colors.transparent;
+    }
+    if (widget.disabledColor != null) {
+      return widget.disabledColor;
+    } else {
+      return color.withOpacity(0.48);
+    }
+  }
+
+  Color getColor() {
+    if (widget.type == GFButtonType.transparent ||
+        widget.type == GFButtonType.outline ||
+        widget.type == GFButtonType.outline2x) {
+      return Colors.transparent;
+    } else {
+      return color;
+    }
+  }
+
+  Color getIconColor() {
+    if (widget.type == GFButtonType.transparent ||
+        widget.type == GFButtonType.outline ||
+        widget.type == GFButtonType.outline2x) {
+      return widget.onPressed != null
+          ? color == GFColors.TRANSPARENT ? GFColors.DARK : color
+          : color.withOpacity(0.48);
+    } else if (color == GFColors.TRANSPARENT) {
+      return widget.onPressed != null ? GFColors.DARK : GFColors.WHITE;
+    } else {
+      return GFColors.WHITE;
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     assert(debugCheckHasMaterial(context));
-
-    Color getBorderColor() {
-      if (widget.onPressed != null) {
-        return this.color;
-      } else {
-        if (widget.disabledColor != null)
-          return widget.disabledColor;
-        else {
-          return this.color.withOpacity(0.48);
-        }
-      }
-    }
-
-    Color getDisabledFillColor() {
-      if (widget.type == GFType.transparent ||
-          widget.type == GFType.outline ||
-          widget.type == GFType.outline2x) return Colors.transparent;
-      if (widget.disabledColor != null)
-        return widget.disabledColor;
-      else {
-        return this.color.withOpacity(0.48);
-      }
-    }
-
-    Color getColor() {
-      if (widget.type == GFType.transparent ||
-          widget.type == GFType.outline ||
-          widget.type == GFType.outline2x)
-        return Colors.transparent;
-      else {
-        return this.color;
-      }
-    }
 
     final Color themeColor =
         Theme.of(context).colorScheme.onSurface.withOpacity(0.12);
@@ -173,81 +182,70 @@ class _GFIconButtonState extends State<GFIconButton> {
           ? getBorderColor()
           : widget.borderSide.color,
       width: widget.borderSide?.width == null
-          ? widget.type == GFType.outline2x ? 2.0 : 1.0
+          ? widget.type == GFButtonType.outline2x ? 2.0 : 1.0
           : widget.borderSide?.width,
     );
 
-    final BorderSide shapeBorder =
-        widget.type == GFType.outline || widget.type == GFType.outline2x
-            ? outlineBorder
-            : widget.borderSide != null
-                ? widget.borderSide
-                : BorderSide(
-                    color: this.color,
-                    width: 0.0,
-                  );
+    final BorderSide shapeBorder = widget.type == GFButtonType.outline ||
+            widget.type == GFButtonType.outline2x
+        ? outlineBorder
+        : widget.borderSide ??
+            BorderSide(
+              color: color,
+              width: 0,
+            );
 
-    ShapeBorder shape;
+    ShapeBorder shapeBorderType;
 
-    if (this.shape == GFIconButtonShape.pills) {
-      shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(20.0), side: shapeBorder);
-    } else if (this.shape == GFIconButtonShape.square) {
-      shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(0.0), side: shapeBorder);
-    } else if (this.shape == GFIconButtonShape.standard) {
-      shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(3.0), side: shapeBorder);
-    } else if (this.shape == GFIconButtonShape.circle) {
-      shape = RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(50.0), side: shapeBorder);
+    if (shape == GFIconButtonShape.pills) {
+      shapeBorderType = RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(20),
+        side: shapeBorder,
+      );
+    } else if (shape == GFIconButtonShape.square) {
+      shapeBorderType = RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(0),
+        side: shapeBorder,
+      );
+    } else if (shape == GFIconButtonShape.standard) {
+      shapeBorderType = RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(3),
+        side: shapeBorder,
+      );
+    } else if (shape == GFIconButtonShape.circle) {
+      shapeBorderType = RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(50),
+        side: shapeBorder,
+      );
     }
 
-    if (widget.size == GFSize.small) {
-      this.height = 30.0;
-      this.width = 30.0;
-      this.iconPixel = 18.0;
-    } else if (widget.size == GFSize.medium) {
-      this.height = 35.0;
-      this.width = 35.0;
-      this.iconPixel = 18.0;
-    } else if (widget.size == GFSize.large) {
-      this.height = 40.0;
-      this.width = 40.0;
-      this.iconPixel = 18.0;
-    }
-
-    getIconColor() {
-      if (widget.type == GFType.transparent ||
-          widget.type == GFType.outline ||
-          widget.type == GFType.outline2x)
-        return widget.onPressed != null
-            ? this.color == getGFColor(GFColor.transparent)
-                ? getGFColor(GFColor.dark)
-                : this.color
-            : this.color.withOpacity(0.48);
-      else if (this.color == getGFColor(GFColor.transparent)) {
-        return widget.onPressed != null
-            ? getGFColor(GFColor.dark)
-            : getGFColor(GFColor.white);
-      } else {
-        return getGFColor(GFColor.white);
-      }
+    if (widget.size == GFSize.SMALL) {
+      height = 30.0;
+      width = 30.0;
+      iconPixel = 18.0;
+    } else if (widget.size == GFSize.MEDIUM) {
+      height = 35.0;
+      width = 35.0;
+      iconPixel = 18.0;
+    } else if (widget.size == GFSize.LARGE) {
+      height = 40.0;
+      width = 40.0;
+      iconPixel = 18.0;
+    } else {
+      height = 35.0;
+      width = 35.0;
+      iconPixel = 18.0;
     }
 
     Widget result = Container(
-      height: widget.shape == GFIconButtonShape.circle
-          ? this.height + 6
-          : this.height,
+      height: widget.shape == GFIconButtonShape.circle ? height + 6.0 : height,
       width: widget.shape == GFIconButtonShape.pills
-          ? this.width + 10
-          : widget.shape == GFIconButtonShape.circle
-              ? this.height + 6
-              : this.width,
+          ? width + 10
+          : widget.shape == GFIconButtonShape.circle ? height + 6 : width,
       padding: widget.padding,
       child: IconTheme.merge(
         data: IconThemeData(
-          size: widget.iconSize > 0.0 ? widget.iconSize : this.iconPixel,
+          size: widget.iconSize > 0.0 ? widget.iconSize : iconPixel,
           color: getIconColor(),
         ),
         child: widget.icon,
@@ -262,7 +260,7 @@ class _GFIconButtonState extends State<GFIconButton> {
     }
 
     BoxDecoration getBoxShadow() {
-      if (widget.type != GFType.transparent) {
+      if (widget.type != GFButtonType.transparent) {
         if (widget.boxShadow == null && widget.buttonBoxShadow != true) {
           return null;
         } else {
@@ -271,28 +269,27 @@ class _GFIconButtonState extends State<GFIconButton> {
                   ? getColor()
                   : getDisabledFillColor(),
               borderRadius: widget.shape == GFIconButtonShape.circle
-                  ? BorderRadius.circular(50.0)
+                  ? BorderRadius.circular(50)
                   : widget.shape == GFIconButtonShape.standard
-                      ? BorderRadius.circular(3.0)
+                      ? BorderRadius.circular(3)
                       : widget.shape == GFIconButtonShape.pills
-                          ? BorderRadius.circular(20.0)
+                          ? BorderRadius.circular(20)
                           : BorderRadius.zero,
               boxShadow: [
                 widget.boxShadow == null && widget.buttonBoxShadow == true
                     ? BoxShadow(
                         color: themeColor,
                         blurRadius: 1.5,
-                        spreadRadius: 2.0,
+                        spreadRadius: 2,
                         offset: Offset.zero,
                       )
-                    : widget.boxShadow != null
-                        ? widget.boxShadow
-                        : BoxShadow(
-                            color: Theme.of(context).canvasColor,
-                            blurRadius: 0.0,
-                            spreadRadius: 0.0,
-                            offset: Offset.zero,
-                          )
+                    : widget.boxShadow ??
+                        BoxShadow(
+                          color: Theme.of(context).canvasColor,
+                          blurRadius: 0,
+                          spreadRadius: 0,
+                          offset: Offset.zero,
+                        )
               ]);
         }
       }
@@ -306,49 +303,42 @@ class _GFIconButtonState extends State<GFIconButton> {
         focusNode: widget.focusNode,
         autofocus: widget.autofocus,
         child: ConstrainedBox(
-          constraints: BoxConstraints(maxWidth: 60.0, maxHeight: 60.0),
+          constraints: const BoxConstraints(maxWidth: 60, maxHeight: 60),
           child: Container(
-            height: widget.shape == GFIconButtonShape.circle
-                ? this.height + 6
-                : this.height,
+            height:
+                widget.shape == GFIconButtonShape.circle ? height + 6 : height,
             width: widget.shape == GFIconButtonShape.pills
-                ? this.width + 10
-                : widget.shape == GFIconButtonShape.circle
-                    ? this.height + 6
-                    : this.width,
-            decoration: widget.type == GFType.solid ? getBoxShadow() : null,
+                ? width + 10
+                : widget.shape == GFIconButtonShape.circle ? height + 6 : width,
+            decoration:
+                widget.type == GFButtonType.solid ? getBoxShadow() : null,
             child: Material(
-              shape: widget.type == GFType.transparent
+              shape: widget.type == GFButtonType.transparent
                   ? null
-                  : widget.borderShape == null ? shape : widget.borderShape,
+                  : widget.borderShape ?? shapeBorderType,
               color: widget.onPressed != null
                   ? getColor()
                   : getDisabledFillColor(),
-              type: widget.type == GFType.transparent
+              type: widget.type == GFButtonType.transparent
                   ? MaterialType.transparency
                   : MaterialType.button,
               child: InkResponse(
                 onTap: widget.onPressed,
                 child: result,
-                focusColor: widget.focusColor != null
-                    ? getGFColor(widget.focusColor)
-                    : Theme.of(context).focusColor,
-                hoverColor: widget.hoverColor != null
-                    ? getGFColor(widget.hoverColor)
-                    : Theme.of(context).hoverColor,
-                highlightColor: widget.highlightColor != null
-                    ? getGFColor(widget.highlightColor)
-                    : Theme.of(context).highlightColor,
-                splashColor: widget.splashColor != null
-                    ? getGFColor(widget.splashColor)
-                    : Theme.of(context).splashColor,
+                focusColor: widget.focusColor ?? Theme.of(context).focusColor,
+                hoverColor: widget.hoverColor ?? Theme.of(context).hoverColor,
+                highlightColor:
+                    widget.highlightColor ?? Theme.of(context).highlightColor,
+                splashColor:
+                    widget.splashColor ?? Theme.of(context).splashColor,
                 radius: math.max(
-                  Material.defaultSplashRadius,
-                  (widget.iconSize +
-                          math.min(widget.padding.horizontal,
-                              widget.padding.vertical)) *
-                      0.7,
-                ),
+                    Material.defaultSplashRadius,
+                    (widget.iconSize +
+                            math.min(
+                              widget.padding.horizontal,
+                              widget.padding.vertical,
+                            )) *
+                        0.7),
               ),
             ),
           ),
