@@ -1,9 +1,9 @@
 /// [GFPosition] is used to position the icon, badges to start or end of the button
-/// See [GFButton] and [GFButtonBadge]
+/// See GFButton and GFButtonBadge
 enum GFPosition {
-  /// [GFPosition.start] is used to place icon, badges to start of the [GFButton] and [GFButtonBadge]
+  /// [GFPosition.start] is used to place icon, badges to start of the GFButton and GFButtonBadge
   start,
 
-  /// [GFPosition.end] is used to place icon, badges to end of the [GFButton] and [GFButtonBadge]
+  /// [GFPosition.end] is used to place icon, badges to end of the GFButton and GFButtonBadge
   end
 }
