@@ -3,12 +3,12 @@ enum GFToggleType {
   /// [GFToggleType.android] is used for basic toggle shape of android mode
   android,
 
-  /// [GFToggleType.custom] is used to change the shape of [GFToggle]
+  /// [GFToggleType.custom] is used to change the shape of GFToggle
   custom,
 
   /// [GFToggleType.ios] is used for basic toggle shape of ios mode
   ios,
 
-  /// [GFToggleType.square] is used for square shape of [GFToggle]
+  /// [GFToggleType.square] is used for square shape of GFToggle
   square
 }
